web/api/requirements.txt
Django==3.2
djangorestframework==3.12.4
django-grappelli
django-nested-admin==3.3.3
django-sortedm2m==3.0.2
django-json-field==0.5.7
django-admin-sortable2
requests
Pillow
sorl-thumbnail==12.7.0
django-object-actions==3.0.2
django-admin-row-actions